% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpqr.R
\name{dpqr-uBS}
\alias{dpqr-uBS}
\alias{duBS}
\alias{puBS}
\alias{quBS}
\alias{ruBS}
\title{The unit-Birnbaum-Saunders distribution}
\usage{
duBS(x, mu, alpha, tau = 0.5, log = FALSE)

puBS(q, mu, alpha, tau = 0.5, lower.tail = TRUE, log.p = FALSE)

quBS(p, mu, alpha, tau = 0.5, lower.tail = TRUE, log.p = FALSE)

ruBS(n, mu, alpha, tau = 0.5)
}
\arguments{
\item{x, q}{vector of positive quantiles.}

\item{mu}{location parameter indicating the \eqn{\tau}-quantile \eqn{\tau \in (0, 1)}.}

\item{alpha}{nonnegative shape parameter.}

\item{tau}{the parameter to specify which quantile is to used.}

\item{log, log.p}{logical; If TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; If TRUE, (default), \eqn{P(X \leq x)} are returned, otherwise \eqn{P(X > x)}.}

\item{p}{vector of probabilities.}

\item{n}{number of observations. If \code{length(n) > 1}, the length is taken to be the number required.}
}
\value{
\code{duBS} gives the density, \code{puBS} gives the distribution function,
\code{quBS} gives the quantile function and \code{ruBS} generates random deviates.

Invalid arguments will return an error message.
}
\description{
Density function, distribution function, quantile function, random taumber generation function
for the unit-Birnbaum-Saunders distribution re-parameterized in terms of the \eqn{\tau}th quantile, \eqn{\tau \in (0,1)}.
}
\examples{
set.seed(6969)
x <- ruBS(n = 1000, mu = 0.5, alpha = 1.5, tau = 0.5)
R <- range(x)
S <- seq(from = R[1], to = R[2], by =  0.01)
hist(x, prob = TRUE, main = 'unit-Weibull distribution')
lines(S, duBS(x = S, mu = 0.5, alpha = 1.5, tau = 0.5), col = 2)
plot(ecdf(x))
lines(S, puBS(q = S, mu = 0.5, alpha = 1.5, tau = 0.5), col = 2)
plot(quantile(x, probs = S), type = "l")
lines(quBS(p = S, mu = 0.5, alpha = 1.5, tau = 0.5), col = 2)

}
\references{
Mazucheli, J., Menezes, A. F. B and Dey, S. (2018). The unit-Birnbaum-Saunders distribution with applications. \emph{Chilean Journal of Statistics}, \bold{9}(1), 47--57.
}
\author{
Josmar Mazucheli

André Felipe B. Menezes
}
