% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hnp.R
\name{hnp}
\alias{hnp}
\alias{hnp.unitBSQuantReg}
\title{Half-Normal Probability Plots with Simulated Envelopes}
\usage{
hnp(object, ...)

\method{hnp}{unitBSQuantReg}(
  object,
  nsim = 99,
  halfnormal = TRUE,
  plot = TRUE,
  level = 0.95,
  resid.type = c("cox-snell", "quantile"),
  ...
)
}
\arguments{
\item{object}{fitted model object of class \code{\link{unitBSQuantReg}}.}

\item{...}{currently not used.}

\item{nsim}{number of simulations used to compute envelope. Default is 99 replications.}

\item{halfnormal}{logical. If \code{TRUE}, a half-normal probability plot is produced. If \code{FALSE}, a normal probability plot is produced.}

\item{plot}{Should the (half-)normal probability plot be plotted? Default is \code{TRUE}.}

\item{level}{confidence level of the simulated envelope. Default is 95\%.}

\item{resid.type}{type of residuals to be used: The currently options are \code{cox-snell} or \code{quantile}.}
}
\description{
Generates (half-)normal probability plots for a \code{\link{unitBSQuantReg}} object.
}
\examples{
set.seed(123)
x <- runif(100, -1, 1)
eta <- 1 + 2 * x
mu <- exp(eta) / (1 + exp(eta))
data <- data.frame(y = ruBS(n = 100, mu, alpha = 1.0, tau = 0.5), x)
fit <- unitBSQuantReg(y ~ x, data = data, tau = 0.5)
hnp(fit, resid.type = "cox-snell")
x1 <- runif(100, -1, 1); x2 <- runif(100, -1, 1)
eta <- 1 + 2 * x1 + 3 * x2
mu <- exp(eta) / (1 + exp(eta))
data <- data.frame(y = ruBS(n = 100, mu, alpha = 1.0, tau = 0.5), x1, x2)
fit <- unitBSQuantReg(y ~ ., data = data, tau = 0.5)
hnp(fit, resid.type = "quantile")

}
\references{
Atkinson, A. C., (1981). Two graphical displays for outlying and influential observations in regression. \emph{Biometrika} \bold{68}(1), 13--20.
}
\author{
Josmar Mazucheli

André Felipe B. Menezes
}
