% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{residuals.unitBSQuantReg}
\alias{residuals.unitBSQuantReg}
\title{Residuals Method for \code{unitBSQuantReg} Objects}
\usage{
\method{residuals}{unitBSQuantReg}(object, type = c("cox-snell", "quantile"), ...)
}
\arguments{
\item{object}{fitted model object of class \code{\link{unitBSQuantReg}}.}

\item{type}{character indicating type of residuals.}

\item{...}{currently not used.}
}
\description{
Extract Cox-Snell and Quantile residuals from a unit-Birnbaum-Saunders object.
}
\references{
Cox, D. R. and Snell, E. J. (1968). A general definition of residuals. \emph{Journal of the Royal Statistical Society. Series B (Methodological)} \bold{2} (30), 248--275.
Dunn, P. K. and Smyth, G. K. (1996). Randomized Quantile Residuals. \emph{Journal of Computational and Graphical Statistics} \bold{4} (3) 236--244.
}
\author{
Josmar Mazucheli
André Felipe B. Menezes
}
