% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitBSQuantReg.R
\name{unitBSQuantReg.control}
\alias{unitBSQuantReg.control}
\title{Control Parameters for the unit-Birnbaum-Saunders quantile regression}
\usage{
unitBSQuantReg.control(method = "BFGS", maxit = 5000, reltol = 1e-08, ...)
}
\arguments{
\item{method}{characters string specifying the method argument passed to \code{optim}.}

\item{maxit}{integer specifying the \code{maxit} argument (maximal number of iterations) passed to \code{optim}.}

\item{reltol}{relative convergence tolerance passed to \code{optim}.}

\item{...}{arguments passed to \code{optim}.}
}
\description{
Parameters that control fitting of the unit-Birnbaum-Saunders regression models using \code{unitBSQuantReg}.
}
\author{
Josmar Mazucheli

André Felipe B. Menezes
}
