% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitBSQuantReg.R
\name{unitBSQuantReg}
\alias{unitBSQuantReg}
\alias{unitBSQuantReg.fit}
\title{The unit-Birnbaum-Saunders quantile regression}
\usage{
unitBSQuantReg(
  formula,
  tau,
  data,
  link,
  link.alpha = NULL,
  start = NULL,
  control = unitBSQuantReg.control()
)

unitBSQuantReg.fit(
  y,
  X,
  Z = NULL,
  tau,
  link,
  link.alpha,
  start,
  control = unitBSQuantReg.control()
)
}
\arguments{
\item{formula}{symbolic description of the quantile model.}

\item{tau}{the quantile(s) to be estimated, number between 0 and 1.}

\item{data}{data.frame contain the variables in the model.}

\item{link}{character specification of the link function in the quantile model. Default is \code{logit}.}

\item{link.alpha}{character specification of the link function in the shape model. Default is \code{log}.}

\item{start}{an optional vector with starting values for all parameters}

\item{control}{a list of control arguments specified via \code{unitBSQuantReg.control}.}

\item{y}{numeric vector of response variable.}

\item{X}{numeric matrix. Regressor matrix for the quantile model.}

\item{Z}{numeric matrix. Regressor matrix for the shape model. Default is constant shape model.}
}
\description{
Fit the unit-Birnbaum-Saunders quantile regression.
}
\details{
The estimated regression coefficients are in logit scale (default)
and estimated shape parameter in log-scale.
}
\examples{
set.seed(123)
data <- data.frame(y = ruBS(n = 100, mu = 0.5, alpha = 1.0, tau = 0.5))
fit <- unitBSQuantReg(y ~ 1, data = data, tau = 0.5)
exp(fit$coefficients[1]) / (1 + exp(fit$coefficients[1])); exp(fit$coefficients[2])

set.seed(123)
x <- runif(100, -1, 1)
eta <- 1 + 2 * x
mu <- exp(eta) / (1 + exp(eta))
data <- data.frame(y = ruBS(n = 100, mu, alpha = 2.0, tau = 0.5), x)
fit <- unitBSQuantReg(y ~ x, data = data, tau = 0.5)

set.seed(123)
x1 <- rnorm(100)
x2 <- rnorm(100)
eta <- 1 + 2 * x1 + 3 * x2
mu <- exp(eta) / (1 + exp(eta))
data <- data.frame(y = ruBS(n = 100, mu, alpha = 1.0, tau = 0.5), x1, x2)
fit <- unitBSQuantReg(y ~ ., data = data, tau = 0.5)

}
\author{
Josmar Mazucheli

André Felipe B. Menezes
}
